update menu items
