+++ .vscode/settings.json
@@ -1,3 +1,4 @@
 {
-  "prettier.jsxSingleQuote": true
+  "prettier.jsxSingleQuote": true,
+  "standard.enable": false
 }
